chore: reorganizing tests
