+++ IntegrationTester/UnitTests/ProxyServiceTests.swift
@@ -1,15 +1,15 @@
 //
-//  ElementServiceTests.swift
-//  IntegrationTesterTests
+//  ProxyServiceTests.swift
+//  UnitTests
 //
-//  Created by Brian Gonzalez on 11/10/22.
+//  Created by Brian Gonzalez on 12/28/22.
 //
 
 import XCTest
 import BasisTheoryElements
 import BasisTheory
 
-final class ElementServiceTests: XCTestCase {
+final class ProxyServiceTests: XCTestCase {
     private final var TIMEOUT_EXPECTATION = 5.0
     
     override func setUpWithError() throws {
@@ -17,72 +17,6 @@ final class ElementServiceTests: XCTestCase {
     }
     
     override func tearDownWithError() throws { }
-    
-    func testTokenizeChecksForApplicationTypeOfPublic() throws {
-        let body: [String: Any] = [
-            "myProp": "myValue"
-        ]
-        
-        let privateApiKey = Configuration.getConfiguration().privateBtApiKey!
-        let tokenizeExpectation = self.expectation(description: "Tokenize")
-        BasisTheoryElements.tokenize(body: body, apiKey: privateApiKey) { data, error in
-            XCTAssertNil(data)
-            XCTAssertEqual(error as! TokenizingError, TokenizingError.applicationTypeNotPublic)
-            
-            tokenizeExpectation.fulfill()
-        }
-        
-        waitForExpectations(timeout: TIMEOUT_EXPECTATION)
-    }
-    
-    func testTokenizeReturnsErrorFromApplicationCheck() throws {
-        let body: [String: Any] = [
-            "myProp": "myValue"
-        ]
-        
-        let tokenizeExpectation = self.expectation(description: "Tokenize")
-        BasisTheoryElements.tokenize(body: body, apiKey: "bad api key") { data, error in
-            XCTAssertNil(data)
-            XCTAssertNotNil(error)
-            
-            tokenizeExpectation.fulfill()
-        }
-        
-        waitForExpectations(timeout: TIMEOUT_EXPECTATION)
-    }
-    
-    func testCreateTokenChecksForApplicationTypeOfPublic() throws {
-        let body = CreateToken(type: "token", data: [
-            "myProp": "myValue"
-        ])
-        
-        let privateApiKey = Configuration.getConfiguration().privateBtApiKey!
-        let tokenizeExpectation = self.expectation(description: "Tokenize")
-        BasisTheoryElements.createToken(body: body, apiKey: privateApiKey) { data, error in
-            XCTAssertNil(data)
-            XCTAssertEqual(error as! TokenizingError, TokenizingError.applicationTypeNotPublic)
-            
-            tokenizeExpectation.fulfill()
-        }
-        
-        waitForExpectations(timeout: TIMEOUT_EXPECTATION)
-    }
-    
-    func testCreateReturnsErrorFromApplicationCheck() throws {
-        let body = CreateToken(type: "token", data: [
-            "myProp": "myValue"
-        ])
-        
-        let tokenizeExpectation = self.expectation(description: "Tokenize")
-        BasisTheoryElements.createToken(body: body, apiKey: "bad api key") { data, error in
-            XCTAssertNil(data)
-            XCTAssertNotNil(error)
-            
-            tokenizeExpectation.fulfill()
-        }
-        
-        waitForExpectations(timeout: TIMEOUT_EXPECTATION)
-    }
     
     func testProxyRequest() throws {
         let privateBtApiKey = Configuration.getConfiguration().privateBtApiKey!
